Carve `TestConstructors` out of the monolith (wala/ML#635).
The 96 constructor tests (`tf.ones`/`zeros`/`eye`/`range`/`fill`/
`one_hot`/`constant`/`convert_to_tensor`/`tf.keras.Input`/`np.*`) move
verbatim; the monolith keeps 878, preserving the 1,040 total.
